refactor(react/components): :truck: Se cambia la nomenclatura para mayor consistencia
