+++ src/react/components/LoginForm/LoginForm.jsx
@@ -1,7 +1,7 @@
 import './LoginForm.css';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
-import { TextField } from '../text-field/text-field';
+import { TextField } from '../TextField/TextField';
 import { UI_TEXT } from './ui/text';
 import { EMAIL_OPTIONS, PASSWORD_OPTIONS } from './utils/register-options';
 import { loginService } from '../../../services';
